Allow keyword expressions with zero arguments:
    obj msg;
    f = (Foo) new;

These are visually similar to Smalltalk unary expressions, but they
have a lower precedence.

I hope that the low precedence, together with the fact that a keyword
expression can't be an lvalue, will make the decision between
"obj.foo" and "obj foo" a no-brainer.  However, the role of functions
-- zero-argument functions in particular -- is becoming even less
clear.