Redesigned stack layout and revised register usage.  Local variable and argument offsets are now simple and consistent.  For the first time under Spike/x86, blocks and varargs play well together.
All MethodContexts are now objects; however, they are always allocated
on the stack.  This means that it falls upon the programmer to
explicitly create a Closure using BlockContext|closure in the upward
funargs case.  Otherwise, undefined behavior results.  'thisContext'
works for the first time under x86; but likewise, it is a bit
dangerous.